Extracting gifts to an array accessed through a factory method
Using Array#last/1, we can DRY up code to a proper abstraction of gifts
and use the day to access and join the right gifts together into the
verse. I am using `day - 1` since I am moving the final gift that's
always included out of the array and into its own line where it will
always be appended with the rest of the gifts.